+++ src/main/gen/de/magynhard/crystal/psi/impl/CrystalAnnotationDefinitionImpl.java
@@ -32,4 +32,10 @@ public class CrystalAnnotationDefinitionImpl extends CrystalStubbedAnnotationDef
     else super.accept(visitor);
   }
 
+  @Override
+  @Nullable
+  public CrystalMacroInterpolation getMacroInterpolation() {
+    return PsiTreeUtil.getChildOfType(this, CrystalMacroInterpolation.class);
+  }
+
 }